extract a new interface
